incluye dificultad

+++ Choque.h
@@ -26,14 +26,14 @@ public:
 	static void choque_lista(lista<Asteroide>& la, float radio); //choque con el radio exterior de destrucción
 	static void choque_lista(lista<Disparo>& ld, lista<Asteroide>& la); //choque entre disparos y asteroides
 	//En los choques se pueden producir bonus, por lo que es necesario incluirlos como argumento
-	static void choque_lista(lista<Disparo>& ld, lista<Asteroide>& la, int &puntos, lista<Vida>& lv); //choque entre disparos y asteroides con puntuación (nuestros disparos)
+	static void choque_lista(lista<Disparo>& ld, lista<Asteroide>& la, long int &puntos, lista<Vida>& lv); //choque entre disparos y asteroides con puntuación (nuestros disparos)
 	static void choque_lista(lista<Asteroide>& la, Nave& n); //choque de la nave con los asteroides
 	static void rebote_lista(lista<Asteroide>& la);//choque de los asteroides entre sí
 	static void rebote_lista(lista<Nave_mala>& ln, Borde& b); //choque de naves con el borde
 	static void rebote_lista(lista<Nave_mala>& ln, lista<Asteroide>& la); //las naves rebotan con los asteroides
 	static void rebote_lista(lista<Nave_mala>& ln); //rebotes entre naves 
 	static bool choque_lista(lista<Disparo>& ld, Nave& n); //choque entre disparos y nave
-	static void choque_lista(lista<Disparo>& ld, lista<Nave_mala>& ln, int &puntos, lista<Vida>& lv); //a las naves les llega un disparo y puntuación
+	static void choque_lista(lista<Disparo>& ld, lista<Nave_mala>& ln, long int &puntos, lista<Vida>& lv); //a las naves les llega un disparo y puntuación
 	static void choque_lista(lista<Vida>& lv, Nave& n); //al coger los bonus
 };
